Filter content to add `srcset` and `sizes`
This changes our approach from one where `srcset` and `sizes` are written to the
image markup in post content to one where the responsive attributes are added on the
front end when the page is being rendered. There is a small performance hit but the
benefit is worth the tradeoff for what is gained. What exactly do we gain? Glad you
asked:

* Automatically extend support to images in posts published before we responsive image
support was available.
* Removes the need to update the markup in TinyMCE when the image size is changed.
* Keeps `sizes` attributes, which should probably be adjusted based on the needs of the
theme, from being stored in the database.
* Is non-destructive if a site needs to change available image sizes and regenerate
thumbnails.

Changes:

* Removes the old JS code for handling markup changes in TinyMCE
* Adds `tevkori_filter_content_images()` as a display filter.
* Adds `_tevkori_filter_content_images_callback()` to process filtered images.
* Adds tests